requirements.txt
Flask
Flask-Cors
paramiko
mysql-connector-python
rsa
requests
pyjwt